Separate Match and MatchSubmission components
